+++ breeze-dsl/src/test/kotlin/com/windea/breezeframework/dsl/markup/XmlDslKtTest.kt
@@ -1,6 +1,6 @@
-package com.windea.breezeframework.dsl
+package com.windea.breezeframework.dsl.markup
 
-import com.windea.breezeframework.dsl.markup.*
+import com.windea.breezeframework.dsl.*
 import kotlin.test.*
 
 //TESTED VERY NICE!
@@ -36,20 +36,20 @@ class XmlDslKtTest {
 				comment("""
 					comment
 					comment
-				""".trimIndent()).wrap().indent()
+				""".trimIndent()) wrap true indent true
 				comment("""
 					comment
 					comment
-				""".trimIndent()).wrap()
+				""".trimIndent()) wrap true
 				comment("""
 					comment
 					comment
-				""".trimIndent()).indent()
-				comment("comment").indent().wrap()
-				comment("comment").wrap()
+				""".trimIndent()) indent true
+				comment("comment") indent true wrap true
+				comment("comment") wrap true
 				"element"{
 					+"text"
-				}.unwrap()
+				} wrap false
 			}
 		}
 		println(xml)