+++ CHANGELOG.md
@@ -94,6 +94,7 @@
 * [X] breeze-data 完善equals、hashcode、toString方法的构建方法。
 * [X] breeze-dsl 更新PumlDsl和PumlStateDiagramDsl中。
 * [X] breeze-text Base64的相关扩展。
+* [X] breeze-dsl 完善各种DSL，基本完成PumlStateDiagramDsl。 
 
 * [ ] breeze-core 提供标准颜色的枚举。（考虑改为常量）
 * [ ] breeze-core Uri构建和编码的扩展。